requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
requests==2.32.3
python-dotenv==1.2.1
